src/api/requirements.txt
flask==2.0.2
python-dotenv
numpy==1.21.2
pandas==1.3.3
scikit-image==0.18.3
scipy==1.7.1
opencv-python-headless==4.5.3.56
Pillow==8.3.2
firebase-admin==5.0.3
requests